Test MessagePump directly, I guess it's a public class
